:recycle: refactor: User의 Role을 내부 클래스에서 분리
