+++ airdnb/src/main/java/com/example/airdnb/dto/user/UserResponse.java
@@ -1,13 +1,7 @@
 package com.example.airdnb.dto.user;
 
+import com.example.airdnb.domain.user.Role;
 import com.example.airdnb.domain.user.User;
-import com.example.airdnb.domain.user.User.Role;
-import jakarta.persistence.Column;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
 
 public record UserResponse(
     Long id,
