+++ airdnb/src/main/java/com/example/airdnb/service/AccommodationService.java
@@ -2,7 +2,7 @@ package com.example.airdnb.service;
 
 import com.example.airdnb.domain.accommodation.Accommodation;
 import com.example.airdnb.domain.user.User;
-import com.example.airdnb.domain.user.User.Role;
+import com.example.airdnb.domain.user.Role;
 import com.example.airdnb.dto.accommodation.AccommodationCreationRequest;
 import com.example.airdnb.dto.accommodation.AccommodationResponse;
 import com.example.airdnb.repository.AccommodationRepository;
